// user-management/backend-neispuo-user-management/src/models/siem-logger/siem-log-event-type.enum.ts
export enum SIEMLogEventType {
    PRIVILEGE_PERMISSIONS_CHANGED = 'privilege_permissions_changed',
    PARENT_CREATED = 'parent_created',
    USER_CREATED = 'user_created',
    USER_UPDATED = 'user_updated',
    USER_DELETED = 'user_deleted',
    STUDENT_DISABLE = 'student_disable',
    SCHOOL_BOOK_ACCESS_CHANGE = 'school_book_access_change',
}